# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "1.5"
  title: SiB Colombia - Catalogue API
  description: API for SiB Colombia Catalogue. This API is used as a backend for consult the information of the species records reported in the SiB Colombia.
  termsOfService: http://swagger.io/terms/
  contact:
    name: SiB Colombia Developer Team
    url: http://www.sibcolombia.net
    email: oduque@humboldt.org.co
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
# during dev, should point to your local machine
host: localhost:8000
...
basePath: /api/v1.5
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /records/{id}/hierarchy/{version}:
    x-swagger-router-controller: hierarchy
    get: 
      description: Returns a version of a Hierarchy element by record id and version
      tags:
        - element
      operationId: getHierarchy
      parameters:
        - name: id
          type: string
          in: path
          required: true
        - name: version
          type: integer
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HierarchyVersion"
        # responses may fall through to errors
        "400":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /records/{id}/hierarchy:
    x-swagger-router-controller: hierarchy
    post:
      description: Save a new version of a hierarchy related with the a recrd with the id
      tags:
        - element
      operationId: postHierarchy
      parameters:
        - name: id
          type: string
          in: path
          required: true
        - name: hierarchyVersion
          description: hierarchyVersion object
          in: body
          required: true
          schema:
            $ref: '#/definitions/HierarchyVersion'
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/ResultResponse"
        # responses may fall through to errors
        "400":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /records/to_review/{id}/hierarchy:
    x-swagger-router-controller: hierarchy
    get: 
      description: Returns a list of Hierarchy version with state "to_review"
      tags:
        - element
      operationId: getToReviewHierarchy
      parameters:
        - name: id
          type: string
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HierarchyVersionList"
        # responses may fall through to errors
        "400":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /records/accepted/{id}/hierarchy:
    x-swagger-router-controller: hierarchy
    get: 
      description: Returns the hierarchy version accepted #**
      tags:
        - element
      operationId: getLastAcceptedHierarchy
      parameters:
        - name: id
          type: string
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HierarchyVersion"
        # responses may fall through to errors
        "400":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /records/accepted/{id}/hierarchy/{version}:
    x-swagger-router-controller: hierarchy
    put: 
      description: Update a HierarchyVersion state to accepted
      tags:
        - element
      operationId: setAcceptedHierarchy
      parameters:
        - name: id
          type: string
          in: path
          required: true
        - name: version
          type: integer
          in: path
          required: true
        - name: user
          description: User object
          in: body
          required: true
          schema:
            $ref: '#/definitions/User'
      responses:
        "200":
          description: Success
          
        # responses may fall through to errors
        "400":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
...
  HierarchyVersionList:
    properties:
      hierarchyList:
        type: array
        items:
          $ref: '#/definitions/HierarchyVersion'
...
  HierarchyVersion:
    required:
      - hierarchy
      - id_user
    properties:
      _id:
        type: string
      id_record: 
        type: string
      created: 
        type: string #****
      id_user:
        type: string
      version:
        type: integer
      state:
        type: string
      element:
        type: string
      hierarchy:
        type: array
        items:
          $ref: '#/definitions/Hierarchy'
...
  Agent:
    properties:
      firstName: 
        type: string
      lastName:
        type: string
      organisation:
        type: string
      position:
        type: string
      address:
        type: string
      city:
        type: string
      state:
        type: string
      country: 
        type: string
      postalCode:
        type: string
      phone:
        type: string
      email:
        type: string
      homepage:
        type: string
      personnelDirectory:
        type: string
      personnelIdentifier :
        type: string
...
  Hierarchy:
    properties:
      ancillaryData:
        type: array
        items:
          $ref: '#/definitions/AncillaryData'
      classification:
        type: string
      recommended::
        type: string
      kingdom:
        type: string
      phylum:
        type: string
      classHierarchy:
        type: string
      order:
        type: string
      family:
        type: string
      genus:
        type: string
      subgenus:
        type: string
      taxonRank:
        type: string
      specificEpithet:
        type: string
      infraspecificEpithet:
        type: string
      higherClassification:
        type: string
      parentTaxon:
        type: string
...
  AncillaryData:
    properties:
      dataType:
        type: string 
      mimeType:
        type: string
      agent:
        type: array
        items:
          $ref: '#/definitions/Agent' 
      created:
        type: string
        pattern: '^\d{4}(-\d\d(-\d\d(T\d\d:\d\d(:\d\d)?(\.\d+)?(([+-]\d\d:\d\d)|Z)?)?)?)?$'
      modified:
        type: string
        pattern: '^\d{4}(-\d\d(-\d\d(T\d\d:\d\d(:\d\d)?(\.\d+)?(([+-]\d\d:\d\d)|Z)?)?)?)?$'
      title:
        type: string
      license:
        type: string
      rights:
        type: string
      rightsHolder:
        type: string
      bibliographicCitation:
        type: string
      audience:
        type: array
        items:
          type: string
      source :
        type: string
      subject:
        type: array
        items:
          type: string
      description:
        type: string
      mediaURL:
        type: array
        items:
          type: string
      thumbnailURL:
        type: string
      location:
        type: string
      geoPoint:
        type: string
      additionalInformation:
        type: string
      dataObject:
        type: string
      reference: 
        type: array
        items:
          $ref: '#/definitions/Reference'      
  Reference:
    properties:
      profile_id:
        type: string
      group_id:
        type: string
      created:
        type: string
        pattern: '^\d{4}(-\d\d(-\d\d(T\d\d:\d\d(:\d\d)?(\.\d+)?(([+-]\d\d:\d\d)|Z)?)?)?)?$'
      last_modified:
        type: string
        pattern: '^\d{4}(-\d\d(-\d\d(T\d\d:\d\d(:\d\d)?(\.\d+)?(([+-]\d\d:\d\d)|Z)?)?)?)?$'
      identifiers:
        type: array
        items:
          type: string
      abstract:
        type: string
      type:
        type: string
      source: 
        type: string
      title:
        type: string
      authors:
        type: array
        items:
          type: string
      year:
        type: string
      volume: 
        type: string
      issue:
        type: string
      pages:
        type: string
      series:
        type: string
      chapter: 
        type: string
      websites:
        type: string
      accessed:
        type: string
      publisher:
        type: string
      address:
        type: string
      city:
        type: string
      edition:
        type: string
      institution:
        type: string
      editors:
        type: array
        items:
          type: string
      keywords:
        type: array
        items:
          type: string
      doi:
        type: string
      isbn:
        type: string
      issn:
        type: string
      link:
        type: string
      recordId:
        type: string
...
  User:
    required:
      - id_user
    properties:
      id_user:
        type: string
        pattern: '^(([^<>()[\]\\.,;:\s@\"]+(\.[^<>()[\]\\.,;:\s@\"]+)*)|(\".+\"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$'
  ResultResponse:
    required:
      - message
    properties:
      message:
        type: string
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string